git-svn-id: http://178.79.136.96/svn@335 40e57a83-7501-4b0a-8fdf-d078260c7e35

+++ BusinessObjects/Users.cs
@@ -560,7 +560,7 @@ namespace Sedogo.BusinessObjects
                 cmd.Parameters.Add("@CreatedByFullName", SqlDbType.NVarChar, 200).Value = m_loggedInUser;
                 cmd.Parameters.Add("@LastUpdatedDate", SqlDbType.DateTime).Value = DateTime.Now;
                 cmd.Parameters.Add("@LastUpdatedByFullName", SqlDbType.NVarChar, 200).Value = m_loggedInUser;
-                cmd.Parameters.Add("@FacebookUserID", SqlDbType.Int).Value = (m_facebookUserID == -1 ? (long?)null : m_facebookUserID);
+                cmd.Parameters.Add("@FacebookUserID", SqlDbType.BigInt).Value = (m_facebookUserID == -1 ? (long?)null : m_facebookUserID);
 
                 SqlParameter paramUserID = cmd.CreateParameter();
                 paramUserID.ParameterName = "@UserID";
@@ -620,7 +620,7 @@ namespace Sedogo.BusinessObjects
                 cmd.Parameters.Add("@ProfileText", SqlDbType.NVarChar, 200).Value = m_profileText;
                 cmd.Parameters.Add("@LastUpdatedDate", SqlDbType.DateTime).Value = DateTime.Now;
                 cmd.Parameters.Add("@LastUpdatedByFullName", SqlDbType.NVarChar, 200).Value = m_loggedInUser;
-                cmd.Parameters.Add("@FacebookUserID", SqlDbType.Int).Value = (m_facebookUserID == -1 ? (long?)null : m_facebookUserID);
+                cmd.Parameters.Add("@FacebookUserID", SqlDbType.BigInt).Value = (m_facebookUserID == -1 ? (long?)null : m_facebookUserID);
 
                 cmd.ExecuteNonQuery();
             }
